+++ changelog.md
@@ -1,5 +1,8 @@
 # Changes
 
+## 5/14/19
+* Nested routes
+
 ## 5/9/19
 * Added API POST route for /new/message